connecting to broker using manager classes

+++ app/src/ConfigurationManager.h
@@ -22,7 +22,7 @@ struct DeviceConfiguration
     //TODO: remove hardcoded mac address
     uint8_t device_mac[6] {0x60, 0x52, 0xD0, 0x06, 0x70, 0x27};
     char device_name[16];
-    int ethernet_pin = 5;
+    uint8_t ethernet_pin = 5;
 };
 
 // Never use a JsonDocument to store the configuration!
